Update DecoderErrorDetection.m
Code updated regarding to the below action point. There are some questions written in the code, will be updated later when the questions are clear. So, it is not in a final shape.
"Expand the code for setting up the network of (e.g. 260 devices), each device is given a fixed steiner system element. When the radio device has data to send, it sends it according to the pattern given in the steiner system element. Take a Steiner system and for each steiner system element and associate a radio device with a RNTI.  Take some random subset of the radio devices and see if the decoding is possible or not."