Create vectorized.m
This vectorized version is now not working correctly. It is because when I set the slotVector with zeros in line 22, algorithm removes the interference free(decoded) users from previous slots as well. Then when I compute again indicatorvec and indicatormat some slots which were not possible to decode at the beginning, became interference free (it should not be, because it is not possible to return back in the time and decode them) and we never have decoder error problem at the end.